+++ tests/rpc/test_balances_multicall.py
@@ -1,11 +1,11 @@
-"""Read token balalances using mutlicall"""
+"""Read token balalances using multicall"""
 import os
 from decimal import Decimal
 
 import pytest
 from web3 import Web3
 
-from eth_defi.balances import fetch_erc20_balances_multicall
+from eth_defi.balances import fetch_erc20_balances_multicall, BalanceFetchFailed
 from eth_defi.provider.broken_provider import get_almost_latest_block_number
 from eth_defi.provider.multi_provider import create_multi_provider_web3
 
@@ -46,3 +46,24 @@ def test_fetch_erc20_balances_multicall(web3):
 
     existing_usdc_balance = balances["0x833589fcd6edb6e08f4c7c32d4f71b54bda02913"]
     assert existing_usdc_balance > Decimal(1.0)
+
+
+def test_fetch_erc20_balances_multicall_failure(web3):
+    """Multicall ERC-20 with a broken token."""
+
+    tokens = {
+        "0x9d247fbc63e4d50b257be939a264d68758b43d04",  # Not  atoken
+    }
+
+    # Velvet vault
+    address = "0x9d247fbc63e4d50b257be939a264d68758b43d04"
+
+    block_number = get_almost_latest_block_number(web3)
+
+    with pytest.raises(BalanceFetchFailed):
+        fetch_erc20_balances_multicall(
+            web3,
+            address,
+            tokens,
+            block_identifier=block_number,
+        )